Работа по замечаниям Андрея и Виталия Разделил auth и user
