+++ resources/js/store/modules/user.js
@@ -1,52 +1,19 @@
 import axios from "axios";
 
 export default {
-    actions: {
-        login({ commit }, data) {
-            axios.post('/api/login' , {
-                email: data.email,
-                password: data.password
-            })
-                .then(response => {
-                    const resp = response.data
-                    localStorage.setItem('user-token', resp.token)
-                    let token = localStorage.getItem('user-token')
-                    localStorage.setItem('user-email', resp.user.email)
-                    data.this.$message({
-                        message: 'Добро пожаловать!',
-                        type: 'success'
-                    })
-                    commit('login', resp)
-                    axios.defaults.headers.common['Authorization'] = 'Bearer ' + token
-                    data.this.$router.push('/')
-                })
-                .catch((error) => {
-                    data.this.$message.error(error.response.data.errors)
-                })
-        },
-        logout({ commit }, data) {
-            axios.get('/api/logout')
-                .then(() => {
-                    commit('clearUserData')
-                })
-                .catch((error) => {
-                    data.this.$message.error(error.response.data.errors)
-                })
-        },
-    },
     mutations: {
         setUserEmail(state, email) {
             state.email = email;
         },
-        login(state, data) {
-            state.token = data.token;
-            state.email = data.user.email;
-        },
         clearUserData(state) {
             localStorage.removeItem('user-token');
             localStorage.removeItem('user-email');
-            state.token = null;
-            state.email = null;
+            state.token = '';
+            state.email = '';
+        },
+        setUserData(state, data) {
+            state.token = data.token;
+            state.email = data.user.email;
         },
     },
     state: {
@@ -57,6 +24,5 @@ export default {
         user(state) {
             return state
         },
-        isAuth() {return !!localStorage.getItem('user-token')},
     },
 }
